pictures
Now the GitHub and RCOS logo are on the bottom of the screen. When they are clicked on they will open up a new tab with the github link to the project and the main rcos page. They need to be moved up higher on the page and preferably next to each other