final changes
Now the login page is up to where I want it to look like. I changed the calendar table to be shorter so there is no scrolling. The logos for RCOS and GitHub are now together and in the bottom left corner. The login event part is also a bit bigger to be more noticable.